Add methods to Graph class
Methods #create_vertex_array, #get_possible_moves, and #create_graph
together result in an array of Vertex instances. Each vertex instance
contains a two-coordinate position instance variable and a
possible_moves instance variable, which contains an array with up to
four two-coordinate elements.

The #find method takes a two-coordinate input and outputs the position
in @vertices that contains the Vertex instance with @position equal to
the two-coordinate input.